changed fileSizeThreshold to long

+++ config.c
@@ -1,4 +1,4 @@
 #include "config.h"
 
-int fileSizeThreshold = 512;
+long fileSizeThreshold = 512;
 int synchronizeRecursively = 0;
